requirements.txt
streamlit
python-dotenv
streamlit_option_menu 
streamlit_chat
africastalking
openai
requests
google-generativeai
langchain
langchain-community
langchain-core
langchain-google-genai
google-generativeai
python-dotenv
tqdm
chromadb
faiss-cpu
pypdf
prophet
plotly
st-social-media-links
pickle
nominatim
altair
pillow
reportlab